+++ lib/shared/keys.dart
@@ -1,3 +1,5 @@
 class PrefKeys {
   static const String hasOnboarded = 'hasOnboarded';
+  static const String themeModeInt = 'themeModeInt';
+  static const String isBlackOrWhiteBackground = 'blackOrWhiteBackground';
 }