+++ lib/shared/mobile_app_shell.dart
@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_speed_dial/flutter_speed_dial.dart';
 import 'package:go_router/go_router.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
+import '../account module/create_account_view.dart';
+import 'responsive.dart';
 import 'route_names.dart';
 import 'user_text.dart';
 
@@ -14,7 +17,7 @@ class MobileAppShell extends StatelessWidget {
   Widget build(BuildContext context) => SafeArea(
         child: Scaffold(
           body: child,
-          floatingActionButton: _buildFAB(_getSelectedIndex(context)),
+          floatingActionButton: _buildFAB(context),
           bottomNavigationBar: NavigationBar(
             animationDuration: const Duration(milliseconds: 500),
             selectedIndex: _getSelectedIndex(context),
@@ -31,9 +34,9 @@ class MobileAppShell extends StatelessWidget {
                 label: UserText.homeNavBarRecords,
               ),
               NavigationDestination(
-                icon: Icon(PhosphorIcons.light.dotsThreeOutline),
-                selectedIcon: Icon(PhosphorIcons.fill.dotsThreeOutline),
-                label: UserText.homeNavBarOther,
+                icon: Icon(PhosphorIcons.light.gearSix),
+                selectedIcon: Icon(PhosphorIcons.fill.gearSix),
+                label: UserText.homeNavBarSettings,
               ),
             ],
           ),
@@ -69,10 +72,62 @@ class MobileAppShell extends StatelessWidget {
     }
   }
 
-  FloatingActionButton? _buildFAB(int index) {
+  Widget? _buildFAB(BuildContext context) {
+    final index = _getSelectedIndex(context);
     if (index < 2) {
-      return FloatingActionButton(onPressed: () {});
+      return SpeedDial(
+        activeIcon: PhosphorIcons.regular.x,
+        icon: PhosphorIcons.regular.plus,
+        overlayOpacity: 0,
+        shape: const RoundedRectangleBorder(
+          borderRadius: BorderRadius.all(Radius.circular(16)),
+        ),
+        spaceBetweenChildren: 4,
+        spacing: 8,
+        children: [
+          _buildSpeedDialChild(
+            context,
+            text: 'Add A Record',
+            iconData: PhosphorIcons.fill.pencilSimple,
+          ),
+          _buildSpeedDialChild(
+            context,
+            text: 'Add An Account',
+            iconData: PhosphorIcons.fill.listPlus,
+            onTap: () async => showDialog(
+              context: context,
+              // barrierDismissible: false,
+              builder: (context) =>
+                  const Dialog.fullscreen(child: CreateAccountView()),
+            ),
+          ),
+        ],
+      );
     }
     return null;
   }
+
+  SpeedDialChild _buildSpeedDialChild(
+    BuildContext context, {
+    required String text,
+    required IconData iconData,
+    Key? key,
+    void Function()? onTap,
+  }) =>
+      SpeedDialChild(
+        key: key,
+        backgroundColor: context.theme().colorScheme.primaryContainer,
+        child: Icon(iconData),
+        onTap: onTap,
+        labelWidget: Padding(
+          padding: const EdgeInsets.all(8),
+          child: DecoratedBox(
+            decoration: context.theme().tooltipTheme.decoration!,
+            child: Padding(
+              padding: const EdgeInsets.all(4),
+              child: Text(text, style: context.theme().tooltipTheme.textStyle),
+            ),
+          ),
+        ),
+      );
 }